fix: decimal lot quantities end-to-end, truncate to 6dp at FlexiBee boundary (#930)
* fix: migrate lot consumption pipeline to decimal, truncate to 6dp at FlexiBee boundary

- Bump Rem.FlexiBeeSDK.Client to 0.1.135 (Amount/AmountIssued now decimal)
- Add NuGet.config with local-flexi-sdk source for local pack resolution
- IngredientRequirement.RequiredAmount: double → decimal
- ConsumptionItem.Amount: double → decimal
- FailedConsumptionItem.Amount: double → decimal
- FefoConsumptionAllocator.AllocationEpsilon: double → decimal literal (0.001m)
- FlexiIngredientRequirementAggregator: explicit (decimal) cast at BoM double→decimal boundary
- FlexiIngredientStockValidator: availableStock cast to decimal for comparison with RequiredAmount
- FlexiManufactureClient: remove now-redundant (double) cast on AdjustedConsumption assignment
- FlexiManufactureDocumentService: add RoundForFlexi(decimal) helper using MidpointRounding.ToZero
  at 6dp; apply to all Amount/AmountIssued DTO fields in consumption and production methods;
  update cost accumulation lines to (double)amount; update comments

* fix: update tests for decimal lot quantities, add 6dp truncation regression test

- Fix 14 compilation errors in FefoConsumptionAllocatorTests: add m suffix to
  RequiredAmount/Amount literals and replace Assert.Equal(double, decimal, precision:)
  with exact decimal assertions
- Fix 3 compilation errors in FlexiIngredientStockValidatorTests: m suffix on RequiredAmount
- Rename rounding tests to *TruncatedToSixDecimals and update assertions to use
  Math.Round(..., 6, MidpointRounding.ToZero) without double cast
- Fix remaining double literal comparisons against decimal Amount fields throughout
  FlexiManufactureClientTests (FEFO, scaling, and direct output tests)
- Add regression test SubmitManufactureAsync_LotAmountWithSixDecimalPlaces_IsNotTruncatedOrExpanded
  to catch the original bug: lot with Amount=6.597599m must pass through without drift
- Add clarifying comments to FlexiIngredientRequirementAggregator (double cast boundary)
  and FlexiManufactureDocumentService (cost rounding vs stock precision)
- Remove NuGet.config with /tmp path; add merge note to csproj; register local feed
  in user-level NuGet config instead

* fix: update InsufficientLotStockFilterTests decimal literals for FailedConsumptionItem.Amount change

+++ backend/src/Anela.Heblo.Application/Features/Manufacture/ErrorFilters/FailedConsumptionItem.cs
@@ -5,7 +5,7 @@ public sealed record FailedConsumptionItem(
     string ProductName,
     string? LotNumber,
     DateOnly? Expiration,
-    double Amount);
+    decimal Amount);
 
 public interface IHasFailedConsumptionItems
 {
